+++ src/main/java/net/neganote/monilabs/mixin/NotifiableEnergyContainerMixin.java
@@ -1,15 +1,20 @@
 package net.neganote.monilabs.mixin;
 
+import com.gregtechceu.gtceu.api.capability.recipe.IO;
+import com.gregtechceu.gtceu.api.capability.recipe.RecipeCapability;
 import com.gregtechceu.gtceu.api.machine.MetaMachine;
 import com.gregtechceu.gtceu.api.machine.TickableSubscription;
-import com.gregtechceu.gtceu.api.machine.trait.MachineTrait;
 import com.gregtechceu.gtceu.api.machine.trait.NotifiableEnergyContainer;
+import com.gregtechceu.gtceu.api.machine.trait.NotifiableRecipeHandlerTrait;
+import com.gregtechceu.gtceu.api.recipe.GTRecipe;
+import com.gregtechceu.gtceu.api.recipe.ingredient.EnergyStack;
 
 import com.lowdragmc.lowdraglib.syncdata.field.ManagedFieldHolder;
 
 import net.minecraft.server.level.ServerLevel;
 import net.neganote.monilabs.common.machine.multiblock.CreativeEnergyMultiMachine;
 
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -17,10 +22,11 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.List;
 import java.util.UUID;
 
 @Mixin(value = NotifiableEnergyContainer.class, remap = false)
-public class NotifiableEnergyContainerMixin extends MachineTrait {
+public class NotifiableEnergyContainerMixin extends NotifiableRecipeHandlerTrait<EnergyStack> {
 
     @Shadow
     protected @Nullable TickableSubscription outputSubs;
@@ -49,6 +55,7 @@ public class NotifiableEnergyContainerMixin extends MachineTrait {
             outputSubs = machine.subscribeServerTick(this.outputSubs, this::serverTick);
             UUID uuid = machine.getOwnerUUID();
             if (uuid != null && CreativeEnergyMultiMachine.isCreativeEnergyEnabledFor(uuid)) {
+                notifyListeners();
                 // return 1 less so active transformers won't turn off
                 cir.setReturnValue(getEnergyCapacity() - 1);
             }
@@ -64,6 +71,7 @@ public class NotifiableEnergyContainerMixin extends MachineTrait {
             outputSubs = machine.subscribeServerTick(this.outputSubs, this::serverTick);
             UUID uuid = machine.getOwnerUUID();
             if (uuid != null && CreativeEnergyMultiMachine.isCreativeEnergyEnabledFor(uuid)) {
+                notifyListeners();
                 cir.setReturnValue(energyToAdd);
             }
         }
@@ -72,5 +80,35 @@ public class NotifiableEnergyContainerMixin extends MachineTrait {
     @Override
     public ManagedFieldHolder getFieldHolder() {
         return NotifiableEnergyContainer.MANAGED_FIELD_HOLDER;
+    }
+
+    @Override
+    @Shadow
+    public IO getHandlerIO() {
+        return null;
+    }
+
+    @Override
+    @Shadow
+    public List<EnergyStack> handleRecipeInner(IO io, GTRecipe recipe, List left, boolean b) {
+        return List.of();
+    }
+
+    @Override
+    @Shadow
+    public @NotNull List<Object> getContents() {
+        return List.of();
+    }
+
+    @Override
+    @Shadow
+    public double getTotalContentAmount() {
+        return 0;
+    }
+
+    @Override
+    @Shadow
+    public RecipeCapability<EnergyStack> getCapability() {
+        return null;
     }
 }