+++ changelog.md
@@ -1,3 +1,5 @@
-# Moni Labs 0.21.6
+# Moni Labs 0.21.8
 
-* HOTFIX: stop client from crashing on startup
+* Fix rare error from the TES shader asking for a texture slot it can't get (thanks JenyaRostov)
+* Add auto-complete card support for extended pattern providers (thanks tamer bayar)
+* Add internal static var for an overclocks-as-parallels recipe modifier